hack/auth/pyproject.toml
[project]
name = "auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.11",
    "python-jose[cryptography]>=3.4.0",
    "uvicorn>=0.34.0",
]
